Change fqvm_preferences>label to demand halted vm
_qvm-prefs_ lets you change the label of a qube while it is running.
This is absolutely fine. However, these changes will not
affect window borders until the corresponding qube is rebooted.
With this commit, we will no longer let users change labels
of running qubes, so as to prevent _qmenu_ from behaving in a non
coherent way, regarding the rest of the system.